// src/app/technician/model/technician-menu-data-struct.ts
export class TechnicianMenuDataStruct {
    enableToBranch: boolean;
    legalEntityId: number;
    legalEntityMenuId: number;
    menuName: string;
    menuParamId: number;
    menuParameterName: string;
    menuParameterPath: string;
    menuPlaceholder: string;
    ngModelPropName: string;
    ngmodelProp: string;
}